feat: extract debugger plugin to standalone package
