+++ packages/debugger/README.md
@@ -1,0 +1,16 @@
+# @univerjs/debugger
+
+[![npm version](https://img.shields.io/npm/v/@univerjs/debugger)](https://npmjs.org/packages/@univerjs/debugger)
+[![license](https://img.shields.io/npm/l/@univerjs/debugger)](https://img.shields.io/npm/l/@univerjs/debugger)
+
+## Introduction
+
+> This plugin provides a lot of utilities to help you debug Univer. **Never use this plugin in your production environment!**
+
+## Usage
+
+### Installation
+
+```shell
+npm i @univerjs/debugger
+```
