+++ packages/debugger/src/commands/operations/confirm.operation.ts
