+++ packages/debugger/src/commands/operations/sidebar.operation.ts
