+++ packages/debugger/src/components/VueI18nIcon.vue
