+++ packages/debugger/tsconfig.node.json
@@ -1,0 +1,4 @@
+{
+    "extends": "@univerjs/shared/tsconfigs/node",
+    "include": ["vite.config.ts"]
+}
